refitting now reuses the fitter and is only kept if better then origenal fit
