+++ tsadar/utils/process/postprocess.py
@@ -1,15 +1,16 @@
 from typing import Dict
 from collections import defaultdict
+from flatten_dict import flatten, unflatten
 
 import time, tempfile, mlflow, os, copy
 
 import numpy as np
-import scipy.optimize as spopt
-import equinox as eq
+import jax
 
 from tsadar.utils.plotting import plotters
 from tsadar.inverse.loss_function import LossFunction
-from tsadar.core.modules import get_filter_spec
+from tsadar.core.modules import IonParams
+from tsadar.inverse.loops import one_d_loop
 
 
 def recalculate_with_chosen_weights(
@@ -258,59 +259,64 @@ def refit_bad_fits(config, batch_indices, all_data, loss_fn, sa, fitted_weights,
 
     red_losses_init = losses_init / (1.1 * reduced_points)
     true_batch_size = config["optimizer"]["batch_size"]
-    # config["optimizer"]["batch_size"] = 1
+
     mlflow.log_metrics({"number of fits": len(batch_indices.flatten())})
     mlflow.log_metrics({"number of refits": int(np.sum(red_losses_init > config["other"]["refit_thresh"]))})
+
+    sample_indices = np.arange(max(len(all_data["e_data"]), len(all_data["i_data"])))
 
     for i in batch_indices.flatten()[red_losses_init > config["other"]["refit_thresh"]]:
         if i == 0:
             continue
-
-        batch = {
-            "e_data": np.reshape(all_data["e_data"][i], (1, -1)),
-            "e_amps": np.reshape(all_data["e_amps"][i], (1, -1)),
-            "i_data": np.reshape(all_data["i_data"][i], (1, -1)),
-            "i_amps": np.reshape(all_data["i_amps"][i], (1, -1)),
-            "noise_e": np.reshape(all_data["noiseE"][i], (1, -1)),
-            "noise_i": np.reshape(all_data["noiseI"][i], (1, -1)),
-        }
-
-        # previous_weights = {}
-        temp_cfg = copy.copy(config)
+        
+        temp_cfg = copy.deepcopy(config)
         temp_cfg["optimizer"]["batch_size"] = 1
-        for species in fitted_weights[(i - 1) // true_batch_size].keys():
-            for key in fitted_weights[(i - 1) // true_batch_size][species].keys():
-                if config["parameters"][species][key]["active"]:
-                    temp_cfg["parameters"][species][key]["val"] = float(
-                        fitted_weights[(i - 1) // true_batch_size][species][key][(i - 1) % true_batch_size]
-                    )
+        
+        def func(x):
+            #i, true_batch_size
+            if hasattr(x, '__len__'):
+                return {'val': x[(i - 1) % true_batch_size]}
+            else:
+                return {'val': x}
+        
+        def extract(x):
+            #i, true_batch_size would idealy be inputs but i cant figure out how to pass variables
+            if isinstance(x,list) or len(np.shape(x))>0:
+                return x[(i - 1) % true_batch_size]
+            else:
+                return x
+            
+        def insert(x,y):
+            #i, true_batch_size
+            if isinstance(x,list):
+                x[i % true_batch_size] = y[0]
+                return x
+            elif len(np.shape(x))>0:
+                x = x.at[i % true_batch_size].set(y[0])
+                return x
+            else:
+                return y
+            
+        prev_weights = fitted_weights[(i - 1) // true_batch_size]
+        prev_weights = jax.tree.map(extract,prev_weights,
+                                     is_leaf= lambda x: isinstance(x,list) and not isinstance(x[0], IonParams))
+        prev_weights = prev_weights.get_unnormed_params()
+        prev_weights = jax.tree.map(lambda x: {'val': x}, prev_weights)
+        prev_weights['electron']['fe']={'m': prev_weights['electron']['m']}
+        del prev_weights['electron']['m']
+        
+        temp_params = flatten(temp_cfg["parameters"])
+        temp_params.update(flatten(prev_weights))
+        temp_cfg["parameters"] = unflatten(temp_params)
+        #temp_cfg["parameters"] = temp_cfg["parameters"] | prev_weights 
+        new_weights, overall_loss, _ = one_d_loop(temp_cfg, all_data, sa, sample_indices, 1)
 
-        loss_fn_refit = LossFunction(temp_cfg, sa, batch)
-
-        # loss_fn_refit.flattened_weights, loss_fn_refit.unravel_pytree = ravel_pytree(previous_weights)
-
-        res = spopt.minimize(
-            loss_fn_refit.vg_loss if config["optimizer"]["grad_method"] == "AD" else loss_fn_refit.loss,
-            np.copy(loss_fn_refit.flattened_weights),
-            args=batch,
-            method=config["optimizer"]["method"],
-            jac=True if config["optimizer"]["grad_method"] == "AD" else False,
-            bounds=loss_fn_refit.bounds,
-            options={"disp": True, "maxiter": config["optimizer"]["num_epochs"]},
-        )
-        cur_result = loss_fn_refit.unravel_pytree(res["x"])
-
-        if res < losses_init[i]:
-            for species in cur_result.keys():
-                for key in cur_result[species].keys():
-                    fitted_weights[i // true_batch_size][species][key] = (
-                        fitted_weights[i // true_batch_size][species][key]
-                        .at[i % true_batch_size]
-                        .set(cur_result[species][key][0])
-                    )
-        return losses_init
-
-    config["optimizer"]["batch_size"] = true_batch_size
+        if overall_loss < losses_init[i]:
+            fitted_weights[(i - 1) // true_batch_size] = jax.tree.map(insert,
+                                    fitted_weights[(i - 1) // true_batch_size], 
+                                    new_weights[0],
+                                    is_leaf= lambda x: isinstance(x,list) and not isinstance(x[0], IonParams))
+    return losses_init
 
 
 def process_data(config, sample_indices, all_data, all_axes, loss_fn, fitted_weights, losses_init, used_points, t1, td):
@@ -320,7 +326,7 @@ def process_data(config, sample_indices, all_data, all_axes, loss_fn, fitted_wei
 
     reduced_points = (used_points - num_params)*config["optimizer"]["batch_size"]
 
-    if losses_init == []:
+    if len(losses_init) == 0:
         losses_init = losses
     mlflow.log_metrics({"postprocessing time": round(time.time() - t1, 2)})
     mlflow.set_tag("status", "plotting")
